requirements.txt
autopep8==1.6.0
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
mysqlclient==2.1.0